requirements.txt
twilio==8.7.0
google-generativeai==0.5.1
openai-whisper==20231117
setuptools-rust==1.9.0
fastapi==0.130.0
uvicorn==0.23.2
python-multipart==0.0.6
